requirements.txt
aioredis==1.3.1
asgiref==3.2.3
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.3.1
Automat==20.2.0
boto3==1.14.16
botocore==1.17.16
cffi==1.14.0
channels==2.4.0
channels-redis==2.4.2
click==7.1.1
constantly==15.1.0
cryptography==2.8
daphne==2.4.1
dash==1.9.1
dash-core-components==1.8.1
dash-html-components==1.0.2
dash-renderer==1.2.4
dash-table==4.6.1
dj-database-url==0.5.0
Django==3.0.3
django-crispy-forms==1.9.0
django-heroku==0.3.1
django-plotly-dash==1.1.5
django-redis==4.11.0
django-storages==1.9.1
docutils==0.15.2
dpd-components==0.1.0
Flask==1.1.1
Flask-Compress==1.4.0
future==0.18.2
gunicorn==20.0.4
hiredis==1.0.1
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
itsdangerous==1.1.0
Jinja2==2.11.1
jmespath==0.10.0
MarkupSafe==1.1.1
msgpack==0.6.2
Pillow==7.0.0
plotly==4.5.4
psycopg2==2.8.5
psycopg2-binary==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
python-dateutil==2.8.1
pytz==2019.3
redis==3.4.1
retrying==1.3.3
s3transfer==0.3.3
service-identity==18.1.0
six==1.14.0
sqlparse==0.3.0
Twisted==20.3.0
txaio==20.1.1
urllib3==1.25.9
Werkzeug==1.0.0
whitenoise==5.1.0
zope.interface==5.0.1
